Update Gemfile.lock with new development dependencies

+++ test/test_migrations.rb
@@ -1,34 +1,85 @@
+# frozen_string_literal: true
+
 require "test_helper"
 require "migrations/validator"
 require "tempfile"
 require "pathname"
+require "fileutils"
 
 module Migrations
   class TestValidator < Minitest::Test
     def setup
       @temp_dir = Pathname.new(Dir.mktmpdir)
-      @validator = Validator.new(@temp_dir)
+      @migrations_dir = @temp_dir.join("db/migrate")
+      FileUtils.mkdir_p(@migrations_dir)
+      @validator = Validator.new(@migrations_dir)
     end
 
     def teardown
-      FileUtils.remove_entry(@temp_dir)
+      FileUtils.remove_entry(@temp_dir) if @temp_dir&.exist?
     end
 
     def test_validates_directory_exists
+      FileUtils.remove_entry(@migrations_dir)
       assert_raises(Error) do
         @validator.validate
       end
     end
 
     def test_validates_migration_files
-      FileUtils.mkdir_p(@temp_dir)
-      
-      # Create a valid migration
-      create_migration("20240101000000_create_users.rb", <<~RUBY)
+      create_test_migrations
+
+      assert_empty @validator.errors
+    end
+
+    def test_detects_missing_down_method
+      create_test_migrations_without_down
+
+      assert_includes @validator.errors, "20240101000000_create_users.rb: Missing down method for rollback"
+    end
+
+    def test_detects_missing_timestamps
+      create_test_migrations_without_timestamps
+
+      assert_includes @validator.errors, "20240101000000_create_users.rb: Missing timestamps"
+    end
+
+    def test_detects_missing_foreign_keys
+      create_test_migrations_without_foreign_keys
+
+      assert_includes @validator.errors, "20240101000001_create_posts.rb: Missing foreign key for author_id"
+    end
+
+    private
+
+    def create_test_migrations
+      create_users_migration
+      create_posts_migration
+    end
+
+    def create_test_migrations_without_down
+      create_users_migration_without_down
+      create_posts_migration
+    end
+
+    def create_test_migrations_without_timestamps
+      create_users_migration_without_timestamps
+      create_posts_migration
+    end
+
+    def create_test_migrations_without_foreign_keys
+      create_users_migration
+      create_posts_migration_without_foreign_keys
+    end
+
+    def create_users_migration
+      content = <<~RUBY
         class CreateUsers < ActiveRecord::Migration[7.0]
-          def up
+          def change
             create_table :users do |t|
               t.string :name
+              t.string :email
+              t.timestamps
             end
           end
 
@@ -37,31 +88,86 @@ module Migrations
           end
         end
       RUBY
+      write_migration("20240101000000_create_users.rb", content)
+    end
 
-      # Should not raise an error for valid migration
-      assert_nothing_raised do
-        @validator.validate
-      end
-
-      # Create an invalid migration (missing down method)
-      create_migration("20240101000001_add_email_to_users.rb", <<~RUBY)
-        class AddEmailToUsers < ActiveRecord::Migration[7.0]
-          def up
-            add_column :users, :email, :string
+    def create_users_migration_without_down
+      content = <<~RUBY
+        class CreateUsers < ActiveRecord::Migration[7.0]
+          def change
+            create_table :users do |t|
+              t.string :name
+              t.string :email
+              t.timestamps
+            end
           end
         end
       RUBY
-
-      # Should raise an error for invalid migration
-      assert_raises(Error) do
-        @validator.validate
-      end
+      write_migration("20240101000000_create_users.rb", content)
     end
 
-    private
+    def create_users_migration_without_timestamps
+      content = <<~RUBY
+        class CreateUsers < ActiveRecord::Migration[7.0]
+          def change
+            create_table :users do |t|
+              t.string :name
+              t.string :email
+            end
+          end
 
-    def create_migration(filename, content)
-      File.write(@temp_dir.join(filename), content)
+          def down
+            drop_table :users
+          end
+        end
+      RUBY
+      write_migration("20240101000000_create_users.rb", content)
+    end
+
+    def create_posts_migration
+      content = <<~RUBY
+        class CreatePosts < ActiveRecord::Migration[7.0]
+          def change
+            create_table :posts do |t|
+              t.string :title
+              t.text :content
+              t.references :author, foreign_key: { to_table: :users }
+              t.timestamps
+            end
+          end
+
+          def down
+            drop_table :posts
+          end
+        end
+      RUBY
+      write_migration("20240101000001_create_posts.rb", content)
+    end
+
+    def create_posts_migration_without_foreign_keys
+      content = <<~RUBY
+        class CreatePosts < ActiveRecord::Migration[7.0]
+          def change
+            create_table :posts do |t|
+              t.string :title
+              t.text :content
+              t.integer :author_id
+              t.timestamps
+            end
+          end
+
+          def down
+            drop_table :posts
+          end
+        end
+      RUBY
+      write_migration("20240101000001_create_posts.rb", content)
+    end
+
+    def write_migration(filename, content)
+      file = @migrations_dir.join(filename)
+      FileUtils.mkdir_p(@migrations_dir)
+      File.write(file, content)
     end
   end
-end 
+end
